Release/v4.0.0

+++ registry-core-api/src/main/java/gr/uoa/di/madgik/registry/domain/BatchResult.java
@@ -1,3 +1,19 @@
+/**
+ * Copyright 2018-2025 OpenAIRE AMKE & Athena Research and Innovation Center
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package gr.uoa.di.madgik.registry.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -11,13 +27,13 @@ public class BatchResult {
 
     private String status;
 
-    private int readCount;
+    private long readCount;
 
-    private int writeCount;
+    private long writeCount;
 
-    private int readSkipCount;
+    private long readSkipCount;
 
-    private int writeSkipCount;
+    private long writeSkipCount;
 
     public boolean isDroped() {
         return droped;
@@ -35,35 +51,35 @@ public class BatchResult {
         this.status = status;
     }
 
-    public int getReadCount() {
+    public long getReadCount() {
         return readCount;
     }
 
-    public void setReadCount(int readCount) {
+    public void setReadCount(long readCount) {
         this.readCount = readCount;
     }
 
-    public int getWriteCount() {
+    public long getWriteCount() {
         return writeCount;
     }
 
-    public void setWriteCount(int writeCount) {
+    public void setWriteCount(long writeCount) {
         this.writeCount = writeCount;
     }
 
-    public int getReadSkipCount() {
+    public long getReadSkipCount() {
         return readSkipCount;
     }
 
-    public void setReadSkipCount(int readSkipCount) {
+    public void setReadSkipCount(long readSkipCount) {
         this.readSkipCount = readSkipCount;
     }
 
-    public int getWriteSkipCount() {
+    public long getWriteSkipCount() {
         return writeSkipCount;
     }
 
-    public void setWriteSkipCount(int writeSkipCount) {
+    public void setWriteSkipCount(long writeSkipCount) {
         this.writeSkipCount = writeSkipCount;
     }
